Move pay-fee code into a shared controller
